requirements.txt
aiogram
gspread
oauth2client
qrcode
aiohttp
python-dotenv